Completed the implementation of the Fraction struct
Fraction is now passing all of its test cases.
Fraction is still having trouble with equality comparisons when converting from a floating point number to a Fraction. This happens when the Fraction represents a repeating decimal such as 1/3. This is due to an insufficient way of passing this data to the Fraction, not from the class itself.
Can potentially work out this problem, but would be easier at this point to sanitize the input rather than implement a workaround for this. The Fraction struct is not meant for external use and this should suffice.